add action for taking x with y
This is mainly a blocking action. There's no real way to make
state change in the game using it. But it should be possible to
try, because the game itself accidentally signposted that the
glass shards must be useful. Sigh.

So I changed that message, too. Just to make it clearer.